extensions: added MPRIS extension.
This extension is an Exaile-extension clone that supports every Media Player exporting their interfaces using the MPRIS specfication.
As in Semplice 7 we will ship Pragha instead of Exaile, this is now the default music extension.
Other media players tested are Spotify and Rhythmbox.

Note that Exaile doesn't work even with the MPRIS plugin enabled, as it is based on an old version of the MPRIS specification.